before/scene_processing/requirements.txt

google-cloud-videointelligence==2.12.0
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
python-ffmpeg==2.0.10